+++ src/GitHub_Updater/Base.php
@@ -607,41 +607,6 @@ class Base {
 	}
 
 	/**
-	 * Used to set_site_transient and checks/stores transient id in array.
-	 *
-	 * @param $id
-	 * @param $response
-	 *
-	 * @return bool
-	 */
-	protected function set_transient( $id, $response ) {
-		$repo                  = isset( $this->type ) ? $this->type->repo : 'ghu';
-		$transient             = 'ghu-' . md5( $repo );
-		$this->response[ $id ] = $response;
-		if ( ! in_array( $transient, self::$transients, true ) ) {
-			self::$transients[] = $transient;
-		}
-		set_site_transient( $transient, $this->response, ( self::$hours * HOUR_IN_SECONDS ) );
-
-		return true;
-	}
-
-	/**
-	 * Returns site_transient and checks/stores transient id in array.
-	 *
-	 * @return array
-	 */
-	protected function get_transient() {
-		$repo      = isset( $this->type->repo ) ? $this->type->repo : 'ghu';
-		$transient = 'ghu-' . md5( $repo );
-		if ( ! in_array( $transient, self::$transients, true ) ) {
-			self::$transients[] = $transient;
-		}
-
-		return get_site_transient( $transient );
-	}
-
-	/**
 	 * Delete all transients from array of transient ids.
 	 *
 	 * @param $type